rename HFInvMassFitter files according to O2Physics
